server/requirements.txt
django>=5.1
djangorestframework>=3.15
drf-spectacular>=0.27
psycopg2-binary>=2.9
Pillow>=10.0
djangorestframework-simplejwt>=5.3
celery[redis]>=5.4
redis>=5.0
httpx>=0.27
